+++ lib/dolly/document.rb
@@ -1,8 +1,7 @@
 require "dolly/query"
 require "dolly/property"
 require 'dolly/timestamps'
-require "dolly/mango_query"
-require 'dolly/scope'
+require "dolly/mango"
 
 module Dolly
   class Document
@@ -155,11 +154,11 @@ module Dolly
       def mango_scope scope_name, scope
         self.mango_scopes ||= {}
         name = scope_name.to_sym
-        self.mango_scopes[name] = lambda { |query_object, args| Dolly::Scope.new(query_object, scope, args)}
+        self.mango_scopes[name] = lambda { |query_object, args| Dolly::Mango::Scope.new(query_object, scope, args)}
 
         (class << self; self end).instance_eval do
           define_method name do |*args|
-            self.mango_scopes[name].call(Dolly::MangoQuery.new(self), *args)
+            self.mango_scopes[name].call(Dolly::Mango::Query.new(self), *args)
           end
         end
       end
