+++ lib/dolly/mango.rb
@@ -1,5 +1,7 @@
 require 'dolly/mango/query_validator'
 require 'dolly/mango/selector'
+require 'dolly/mango/query'
+require 'dolly/mango/scope'
 
 module Dolly
   module Mango
